+++ src/quacc/settings.py
@@ -406,16 +406,13 @@ class QuaccSettings(BaseSettings):
     )
 
     # ---------------------------
-    # Debug Settings
+    # Logger Settings
     # ---------------------------
-    DEBUG: bool = Field(
-        False,
-        description=(
-            """
-            Whether to run in debug mode. This will set the logging level to DEBUG,
-            ASE logs (e.g. optimizations, vibrations, thermo) are printed to stdout.
-            """
-        ),
+    LOG_FILENAME: Optional[Path] = Field(
+        None, description="Path to store the log file."
+    )
+    LOG_LEVEL: Optional[Literal["CRITICAL", "ERROR", "WARNING", "INFO", "DEBUG"]] = (
+        Field("INFO", description=("Logger level."))
     )
 
     # --8<-- [end:settings]
@@ -467,12 +464,9 @@ class QuaccSettings(BaseSettings):
         cls, v: Union[str, tuple[str, str]]
     ) -> tuple[str, str]:
         """Clean up Espresso parallel command."""
-        parsl_mpi_prefix = os.environ.get("PARSL_MPI_PREFIX")
 
         if isinstance(v, str):
             v = (v, "")
-        if parsl_mpi_prefix:
-            v = (parsl_mpi_prefix, v[1])
 
         return v