+++ CHANGELOG.md
@@ -4,11 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project generally adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.3]
+
+### Added
+
+- Added support for `preset: Path` type in the `Espresso` calculator
+- Raise a custom `JobFailure` error when a calculation fails, which stores the directory where the files remain as well as the parent error
+
+### Fixed
+
+- Fixed Parsl MPI support for Espresso recipes
+- Fixed miscellaneous inconsistencies with `preset` handling
+
 ## [0.11.2]
 
 ### Added
 
-- Added supprot for `preset: Path` type in the `Vasp` calculator
+- Added support for `preset: Path` type in the `Vasp` calculator
 
 ## [0.11.1]
 
@@ -230,7 +242,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
-- Fixed user setting of logger level
+- Fixed user setting of LOGGER level
 - Fixed a glob-based issue with `copy_decompress_files`
 
 ## [0.7.1]
@@ -445,7 +457,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
-- Raised the default logger level from `WARNING` to `INFO`
+- Raised the default LOGGER level from `WARNING` to `INFO`
 
 ### Fixed